admin check

+++ build/routers/salesman.router.js
@@ -18,30 +18,51 @@ router.route('/').get(async (req, res) => {
     }
 });
 router.route('/').post(async (req, res) => {
-    try {
-        const salesmans = await services_1.salesmanService.addSalesman({ ...req.body });
-        res.json(salesmans);
+    const adminPassword = req.body.adminPassword;
+    if (adminPassword === '12321') {
+        try {
+            const salesmans = await services_1.salesmanService.addSalesman(req.body.dto);
+            res.json(salesmans);
+        }
+        catch (err) {
+            res.json(err);
+        }
     }
-    catch (err) {
-        res.json(err);
+    else {
+        res.statusCode === 403;
+        res.json({ message: 'Incorrect admin password' });
     }
 });
 router.route('/').put(async (req, res) => {
-    try {
-        const salesman = await services_1.salesmanService.updateSalesman(req.body);
-        res.json(salesman);
+    const adminPassword = req.body.adminPassword;
+    if (adminPassword === '12321') {
+        try {
+            const salesman = await services_1.salesmanService.updateSalesman(req.body.dto);
+            res.json(salesman);
+        }
+        catch (err) {
+            return err;
+        }
     }
-    catch (err) {
-        return err;
+    else {
+        res.statusCode === 403;
+        res.json({ message: 'Incorrect admin password' });
     }
 });
 router.route('/').delete(async (req, res) => {
-    try {
-        const id = await services_1.salesmanService.removeSalesman(req.body.salesmanId);
-        res.json({ id, message: 'Deleted' });
+    const adminPassword = req.body.adminPassword;
+    if (adminPassword === '12321') {
+        try {
+            const id = await services_1.salesmanService.removeSalesman(req.body.salesmanId);
+            res.json({ id, message: 'Deleted' });
+        }
+        catch (err) {
+            res.json(err);
+        }
     }
-    catch (err) {
-        res.json(err);
+    else {
+        res.statusCode === 403;
+        res.json({ message: 'Incorrect admin password' });
     }
 });
 exports.salesmanRouter = router;
